完善场景元素

+++ game-server/src/main/java/com/kingston/mmorpg/game/scene/model/Grid.java
@@ -10,11 +10,11 @@ public class Grid {
 	/** 宽度为25像素 */
 	public static final int WIDTH = 25;
 	/** 掩码 */
-	private int code;
+	private byte code;
 
-	private final static int WALK = 1;
-	private final static int SAFE = 1 << 1;
-	private final static int BLOCK = 1 << 2;
+	private final static byte WALK = 1;
+	private final static byte SAFE = 1 << 1;
+	private final static byte BLOCK = 1 << 2;
 
 	public boolean canWalk() {
 		return (code & WALK) > 0;
